new tests moved as a subdirectory of the main tests
